Fixing some errors BMGEditing.py
# Testing BMGEdit

Testing the BMGedit method, when I try to use it, I obtain the error

```python
File "/home/jarr/Documents/Taller/revolutionhtl/hug_free.py", line 31, in build_graph
    solver.build(method, objective)
  File "/home/jarr/Documents/Taller/bmgedit/BMGEditing.py", line 85, in build
    binary_triples=self.binary_triples)
                   ^^^^^^^^^^^^^^^^^^^
AttributeError: 'BMGEditor' object has no attribute 'binary_triples'
```

**The error:** When using the class `Build2` in the lines 77 and 100 of the file `BMGEditing.py`:

It is specified the keyword argument `binary_triples=self.binary_triples`. This is wrong because the object `self.binary_triples` doesn't exists.

The correct attribute where the triples are stored is `self.R`, this is specified in the method `__init__` of the class `BMGEditor`, in the lines 34-42.

**Another error:** Let's look to again at lines 77 and 100 of the file `BMGEditing.py`:

After the correction of the previous error, we use the following keyword argument `binary_triples=self.R` when calling the class `Build2`.

This is still wrong: looking at the method `__init__` of the class `Build2` in the file `Build.py`:

```python
class Build2:
    """BUILD / MTT algorithm with optimal objective partition."""
    
    def __init__(self, R, L, F=None,
                 allow_inconsistency=True,
                 binarize=False,
                 part_method='mincut',
                 obj_function=None,
                 minimize=True,
                 obj_function_args=None,
                 greedy_repeats=5,
                 weighted_mincut=False, triple_weights=None):

[...]
```

Note that there is NOT a keyword argument ` binary_triples`, then the lines 77 and 100 of the file `BMGEditing.py` are wrong.

On the other hand, in the method `__init__` of the class `Build2` in the file `Build.py`, there is an attribute `R` which is designed to hold the triples (no matter if they are classic informative triples, or if they are binary triples).

If you look back in the file `BMGEditing.py`, when using the class `Build2`, the triples are specified in the first argument (lines 77 and 100).